+++ src/api/config/requestUrl.ts
@@ -1,0 +1,18 @@
+export const REQUEST_API = {
+  AUTH: {
+    SIGNUP: '/signup',
+    LOGIN: '/login',
+  },
+  USER: {
+    INDEX: '/user',
+    MEMBER: '/member',
+  },
+  MILESTONE: {
+    INDEX: '/milestone',
+  },
+  LABEL: {
+    INDEX: '/label',
+  },
+};
+
+export const ROOT_API_URL = '/api';
