--- src/api/requestUrl.ts
@@ -1,11 +1,0 @@
-export const REQUEST_API = {
-  AUTH: {
-    SIGNUP: '/signup',
-    LOGIN: '/login',
-  },
-  USER: {
-    INDEX: '/user',
-  },
-};
-
-export const ROOT_API_URL = '';
